requirements.txt
# Deep Learning Framework
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Federated Learning
flwr>=1.6.0
flwr-datasets>=0.0.2

# Computer Vision & Image Processing
opencv-python>=4.8.0
Pillow>=10.0.0
albumentations>=1.3.1

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Visualization & Monitoring
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.65.0

# Model Architecture
timm>=0.9.0
efficientnet-pytorch>=0.7.1

# Utilities
python-dotenv>=1.0.0
tensorboard>=2.13.0
